# server/openapi/swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Traveling Cook Calculator API
  description: |
    API for managing Cook and Run events - a social cooking experience where teams 
    cook different courses at different locations and move between them.

    ## Authentication
    All endpoints require a valid JWT Bearer token. The JWT should contain:
    - `sub`: User ID
    - `role`: User role (admin, organizer, participant)
    - `exp`: Token expiration time

    ## Rate Limiting
    - 1000 requests per hour per authenticated user
    - 100 requests per hour for unauthenticated endpoints

    ## Error Handling
    The API uses conventional HTTP response codes and returns error details in JSON format.
    All error responses follow the same structure with error codes, messages, and optional details.
servers:
  - url: http://0.0.0.0:3000/
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /cook_and_run/{cookAndRunId}/courses:
    get:
      summary: List all courses for a cook and run project
      description: Returns all courses with optional filtering and sorting
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - name: sort
          in: query
          required: false
          description: Sort courses by time or name
          schema:
            type: string
            enum: [time_asc, time_desc]
            default: time_asc
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
                  count:
                    type: integer
                    description: Total number of courses
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Courses
  /cook_and_run/{cookAndRunId}/course/{courseId}:
    post:
      summary: Create course for cook and run project
      description: Creates a new course for the specified cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunCourseCreateRequest"
      responses:
        "201":
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "409":
          description: Course with this ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Courses

    get:
      summary: Get course details
      description: Returns detailed information about a specific course
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Courses

    patch:
      summary: Update course for cook and run project
      description: Updates an existing course with new data
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunCourseUpdateRequest"
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Courses

    delete:
      summary: Delete course for cook and run project
      description: Permanently deletes a course from the cook and run project
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/CourseId"
      responses:
        "204":
          description: Course deleted successfully
        "400":
          description: Cannot delete course that is assigned to teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Courses
...
  /cook_and_run/{cookAndRunId}/team/{teamId}/notes:
    get:
      summary: Get all notes for a team
      description: Returns all notes associated with a team
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [created_asc, created_desc]
            default: created_desc
      responses:
        "200":
          description: List of team notes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Note"
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Notes
  /cook_and_run/{cookAndRunId}/team/{teamId}/note/{noteId}:
    post:
      summary: Create note for team
      description: Creates a new note for the specified team
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
        - $ref: "#/components/parameters/NoteId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunTeamNoteCreateRequest"
      responses:
        "201":
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "409":
          description: Note with this ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Notes

    delete:
      summary: Delete note for team
      description: Permanently deletes a note from the team
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
        - $ref: "#/components/parameters/TeamId"
        - $ref: "#/components/parameters/NoteId"
      responses:
        "204":
          description: Note deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Team Notes
...
  /cook_and_run/{cookAndRunId}/plan:
    get:
      summary: Get complete event plan
      description: |
        Returns the complete plan for the cook and run event, including
        team assignments, routes, and scheduling information
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      responses:
        "200":
          description: Complete event plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Event Planning

    patch:
      summary: Update event plan
      description: |
        Updates the complete plan for the cook and run event.
        This includes team assignments, walking routes, and other logistics.
      parameters:
        - $ref: "#/components/parameters/CookAndRunId"
      requestBody:
        $ref: "#/components/requestBodies/CookAndRunPlanUpdateRequest"
      responses:
        "200":
          description: Event plan updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
      tags:
        - Event Planning

  # Health check and monitoring endpoints
...
components:
  parameters:
    CookAndRunId:
      name: cookAndRunId
      in: path
      required: true
      description: Unique identifier for the cook and run project
      schema:
        type: string
        format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    CourseId:
      name: courseId
      in: path
      required: true
      description: Unique identifier for the course
      schema:
        type: string
        format: uuid
      example: "456e7890-e12b-34c5-d678-901234567890"
    TeamId:
      name: teamId
      in: path
      required: true
      description: Unique identifier for the team
      schema:
        type: string
        format: uuid
      example: "789e0123-e45f-67g8-h901-234567890123"
    NoteId:
      name: noteId
      in: path
      required: true
      description: Unique identifier for the note
      schema:
        type: string
        format: uuid
      example: "012e3456-e78h-90i1-j234-567890123456"
  requestBodies:
...
    CookAndRunCourseCreateRequest:
      description: Request body to create course for a Cook and Run project
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CourseCreateData"
          examples:
            appetizer:
              summary: Appetizer course
              value:
                name: "Appetizers"
                time: "18:00"
            main_course:
              summary: Main course
              value:
                name: "Main Course"
                time: "20:00"
    CookAndRunCourseUpdateRequest:
      description: Request body to update course for a Cook and Run project
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CourseUpdateData"
...
    CookAndRunTeamNoteCreateRequest:
      description: Request body to create a note for a team
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NoteCreateData"
          examples:
            cooking_tip:
              summary: Cooking tip
              value:
                headline: "Vegetarian Options"
                content: "Please prepare extra vegetarian portions as we have several vegetarian guests"
            logistics:
              summary: Logistics note
              value:
                headline: "Parking Information"
                content: "Street parking available. Alternative: underground garage 2 blocks away"
...
    CookAndRunPlanUpdateRequest:
      description: Request body to update the complete event plan
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Plan"
  responses:
    BadRequestError:
      description: Bad request - invalid input parameters or request format
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalid_uuid:
              summary: Invalid UUID format
              value:
                error: "BAD_REQUEST"
                message: "Invalid UUID format in path parameter"
                timestamp: "2024-03-15T10:00:00Z"
            missing_field:
              summary: Missing required field
              value:
                error: "BAD_REQUEST"
                message: "Missing required fields in request body"
                details:
                  - field: "name"
                    message: "Name is required and cannot be empty"
                timestamp: "2024-03-15T10:00:00Z"
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is required"
                timestamp: "2024-03-15T10:00:00Z"
            invalid_token:
              summary: Invalid or expired token
              value:
                error: "UNAUTHORIZED"
                message: "Authentication token is invalid or expired"
                timestamp: "2024-03-15T10:00:00Z"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            project_not_found:
              summary: Cook and Run project not found
              value:
                error: "NOT_FOUND"
                message: "Cook and Run project not found"
                timestamp: "2024-03-15T10:00:00Z"
            team_not_found:
              summary: Team not found
              value:
                error: "NOT_FOUND"
                message: "Team not found in this project"
                timestamp: "2024-03-15T10:00:00Z"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_failed:
              summary: Multiple validation errors
              value:
                error: "VALIDATION_ERROR"
                message: "Request validation failed"
                details:
                  - field: "email"
                    message: "Invalid email format"
                  - field: "phone"
                    message: "Phone number must be in international format"
                  - field: "members"
                    message: "Number of members must be between 1 and 20"
                timestamp: "2024-03-15T10:00:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            server_error:
              summary: Unexpected server error
              value:
                error: "INTERNAL_SERVER_ERROR"
                message: "An unexpected error occurred. Please try again later."
                timestamp: "2024-03-15T10:00:00Z"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed for one or more fields"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that caused the error
              message:
                type: string
                description: Detailed error message for this field
            required:
              - field
              - message
          description: Detailed error information (optional)
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        trace_id:
          type: string
          description: Unique identifier for tracing this error (optional)
      required:
        - error
        - message
        - timestamp
...
    CourseCreateData:
      type: object
      required:
        - name
        - time
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: Name of the course (e.g., "Appetizers", "Main Course")
          example: "Appetizers"
        time:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]"
          description: Time when this course should be served (24-hour format)
          example: "18:00"
    CourseUpdateData:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: Name of the course
        time:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]"
          description: Time when this course should be served
        has_multiple_hosts:
          type: boolean
          description: Set when a course is allow for multiple hosts
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the course
        name:
          type: string
          maxLength: 64
          description: Name of the course
        time:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]"
          description: Time when this course should be served
        has_multiple_hosts:
          type: boolean
          description: Set when a course is allow for multiple hosts
      required:
        - id
        - name
        - time
        - has_multiple_hosts
...
    NoteCreateData:
      type: object
      required:
        - headline
        - content
      properties:
        headline:
          type: string
          minLength: 1
          maxLength: 64
          description: Short headline for the note
          example: "Vegetarian Options"
        content:
          type: string
          minLength: 1
          maxLength: 1000
          description: Detailed content of the note
          example: "Please prepare extra vegetarian portions as we have several vegetarian guests in the group"
...
    Note:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the note
        headline:
          type: string
          maxLength: 64
          description: Short headline for the note
        content:
          type: string
          maxLength: 1000
          description: Detailed content of the note
        created:
          type: string
          format: date-time
          description: When the note was created
      required:
        - id
        - headline
        - content
        - created
...
    Plan:
      type: object
      properties:
        access:
          type: array
          items:
            type: string
            enum: [link, account]
          description: How participants can access this plan
          example: ["link", "account"]
        introduction:
          type: string
          maxLength: 1000
          description: Welcome text and instructions for participants
          example: "Welcome to our Cook & Run event! Follow your personalized plan below."
        hosting_assignments:
          type: array
          items:
            $ref: "#/components/schemas/Hosting"
          description: Complete list of hosting assignments
        walking_paths:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                course_id:
                  type: string
                  format: uuid
                host_team_id:
                  type: string
                  format: uuid
              required:
                - course_id
                - host_team_id
          description: Walking paths for each team (team_id -> ordered list of courses)
      required:
        - access
        - hosting_assignments
        - walking_paths
    Hosting:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for this hosting assignment
        course_id:
          type: string
          format: uuid
          description: Which course this assignment is for
        team_id:
          type: string
          format: uuid
          description: Which team is hosting
        guest_team_ids:
          type: array
          items:
            type: string
            format: uuid
          description: List of team IDs that will be guests at this course
      required:
        - id
        - course_id
        - team_id
        - guest_team_ids
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user authentication information.

        Token should include the following claims:
        - `sub`: User ID
        - `role`: User role (admin, organizer, participant)
        - `exp`: Token expiration time
        - `iss`: Token issuer (api.travelingcook.com)

        Format: `Authorization: Bearer <token>`

        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# Security considerations:
# - Tokens should be refreshed every 24 hours
# - Use HTTPS in production
# - Implement proper token validation and blacklisting
# - Rate limiting is applied per authenticated user
...